+++ ReleaseNotes.md
@@ -5,6 +5,7 @@
 ### ODC common
 Modified: improved error propagation. Use proper error codes and messages.    
 Fixed: reset topology on shutdown. Topology can be activated and stoped multiple times.    
+Added: execution of a sequence of commands in a batch mode. New "--batch" and "--cmds" command line arguments.
 
 
 ## v0.6 (2020-07-16)